+++ codegen/core/src/main/java/software/amazon/smithy/python/codegen/SmithyPythonDependency.java
@@ -22,7 +22,7 @@ public final class SmithyPythonDependency {
      */
     public static final PythonDependency SMITHY_CORE = new PythonDependency(
             "smithy_core",
-            "~=0.4.0",
+            "~=0.5.0",
             Type.DEPENDENCY,
             false);
 
@@ -69,7 +69,7 @@ public final class SmithyPythonDependency {
      */
     public static final PythonDependency SMITHY_AWS_EVENT_STREAM = new PythonDependency(
             "smithy_aws_event_stream",
-            "~=0.2.0",
+            "~=0.3.0",
             Type.DEPENDENCY,
             false);
 
@@ -78,7 +78,7 @@ public final class SmithyPythonDependency {
      */
     public static final PythonDependency SMITHY_AWS_CORE = new PythonDependency(
             "smithy_aws_core",
-            "~=0.5.0",
+            "~=0.6.0",
             Type.DEPENDENCY,
             false);